feat: allow sending certificate files to the iOS simulator (#1465)
Adds support for sending `.cer` and `.pem` files to iOS Simulator
devices.
Closes #486

### How Has This Been Tested: 
- open an iOS device in Radon
- try sending a `.cer` or `.pem` file with a CA Certificate to the
device (either by drag-n-dropping while holding Shift or using the "Send
File" option in the device settings dropdown)
- Safari should open and ask you to install a "security profile"

### How Has This Change Been Documented:
WIP, waiting on #1456

+++ packages/vscode-extension/src/project/deviceSession.ts
@@ -952,16 +952,43 @@ export class DeviceSession implements Disposable {
   }
 
   public async sendFileToDevice(fileName: string, data: ArrayBuffer): Promise<void> {
-    const tempDir = await fs.promises.mkdtemp(os.tmpdir());
+    let canSafelyRemove = true;
+    const tempDir = await this.getTemporaryFilesDirectory();
+    const tempFileLocation = path.join(tempDir, fileName);
     try {
-      const tempFileLocation = path.join(tempDir, fileName);
       await fs.promises.writeFile(tempFileLocation, new Uint8Array(data));
-      await this.sendFile(tempFileLocation);
+      const result = await this.sendFile(tempFileLocation);
+      canSafelyRemove = result.canSafelyRemove;
     } finally {
-      // NOTE: no `await` here, this can safely go in the background
-      fs.promises.rm(tempDir, { recursive: true }).catch((_e) => {
-        /* silence the errors, it's fine */
-      });
+      if (canSafelyRemove) {
+        // NOTE: no need to await this, it can run in the background
+        fs.promises.rm(tempFileLocation, { force: true }).catch((_e) => {
+          // NOTE: we can ignore errors here, as the file might not exist
+        });
+      }
     }
+  }
+
+  private tempDir: string | undefined;
+  /**
+   * Returns the path to a temporary directory, creating it if it does not already exist.
+   * The directory is created using the system's temporary directory and is cleaned up
+   * automatically when the device session is disposed. Subsequent calls return the same directory path.
+   *
+   * @returns {Promise<string>} The path to the temporary directory.
+   */
+  private async getTemporaryFilesDirectory(): Promise<string> {
+    if (this.tempDir === undefined) {
+      const tempDir = await fs.promises.mkdtemp(os.tmpdir());
+      this.tempDir = tempDir;
+      this.disposables.push(
+        new Disposable(() => {
+          fs.promises.rm(tempDir, { recursive: true }).catch((_e) => {
+            /* silence the errors, it's fine */
+          });
+        })
+      );
+    }
+    return this.tempDir;
   }
 }
